Move authentication logic from browser.views.py to db.manager.py
I realized that some of the things we were checking for didn't need to be checked for, since they're really part of postgres permissions. On a few few other non-postgres things (notably, files), I added some checks in manager.py and took them out of browser.views.

This breaks basically all of the functional_tests.test_database_capabilities... and now that needs to be fixed.