Check OAuth requests to the API for needed scopes.
Default is 'read' scope for HEAD, OPTIONS, GET; 'write' for POST, PUT, PATCH, DELETE.

Update the description for those scopes in the authorization screen.

Don't ask the user for authorization later if they've already approved the app in the past.

Don't check for scopes if the request is using session-based auth, like in the API browser.