Make console forget access tokens from previous users.
Don't fetch a new access token on every visit to console if there already is one for the current user.

Clear the access_token hash from the console URL after a successful login. Otherwise page refreshes will try to use the access_token in that hash instead of logging the user in.